refactor(board): const를 common으로 이동, @RequestMapping(/api/v1) 추가, 메서드 이름 수정, dto 패키지 request, response로 나누기
