+++ src/main/java/com/example/newsfeed/image/repository/BoardImageRepository.java
@@ -1,0 +1,9 @@
+package com.example.newsfeed.image.repository;
+
+import com.example.newsfeed.image.entity.BoardImage;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BoardImageRepository extends JpaRepository<BoardImage, Long> {
+}
